Moved TerminalManager to EditorUtils
