+++ CMakeLists.txt
@@ -15,7 +15,7 @@ if(${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_BINARY_DIR})
   message(FATAL_ERROR "In-source builds not allowed. Please make a new directory (called a buld directory) and run CMake from there. You may need to remove CMakeCache.txt.")
 endif()
 
-add_executable(${PROJECT_NAME} src/kilo.cpp src/Editor.hpp src/Editor.cpp src/TerminalManager.hpp src/TerminalManager.cpp)
+add_executable(${PROJECT_NAME} src/kilo.cpp src/Editor.hpp src/Editor.cpp src/EditorUtils.hpp src/EditorUtils.cpp)
 
 target_compile_options(${PROJECT_NAME} PRIVATE -Wall -Wextra -pedantic)
